added route and changed location for controllerfunction addNewEstimation
